pyproject.toml
[tool.poetry]
name = "quickprint"
version = "0.1.0"
description = ""
authors = ["Johann Fridriksson <johann.fridriksson@silabs.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.1"
rich = "^10.7.0"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
